# services/gateway-api/openapi.yaml
openapi: "3.1.0"
info:
  title: Opencomplai Gateway API
  version: 0.1.0-dev
  description: |
    REST API gateway for the Opencomplai compliance platform.
    All routes proxy to the appropriate Python service.
    This spec is the source of truth for the API contract.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose deployment
paths:
...
  /v1/evidence/events:
    post:
      operationId: appendEvidenceEvent
      summary: Append an event to the immutable ledger (non-idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppendEventRequest"
      responses:
        "201":
          description: Event appended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppendEventResponse"
...
  /v1/docs/generate:
    post:
      operationId: generateDocs
      summary: Generate Annex IV technical documentation dossier (idempotent by system_id + commit_ref)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDocsRequest"
      responses:
        "200":
          description: Dossier generated or returned from cache
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateDocsResponse"
...
components:
  schemas:
...
    AppendEventRequest:
      type: object
      required: [event_type, payload]
      properties:
        event_type:
          type: string
        payload:
          type: object
        signer_id:
          type: string
    AppendEventResponse:
      type: object
      required: [event_id, payload_hash, prev_hash]
      properties:
        event_id:
          type: string
        payload_hash:
          type: string
        prev_hash:
          type: string
...
    GenerateDocsRequest:
      type: object
      required: [system_id, commit_ref]
      properties:
        system_id:
          type: string
        commit_ref:
          type: string
    GenerateDocsResponse:
      type: object
      required: [bundle_checksum, status]
      properties:
        bundle_checksum:
          type: string
        status:
          type: string
          enum: [generated, cached, failed]
        duration_ms:
          type: integer
...